x86: Reduce APIC MMIO value variable to 32-bit length
This avoids evaluating the higher 32 bits on writes. We only support
32-bit writes, so we must only look at the lower 32 bit of the input
register.

+++ hypervisor/arch/x86/apic.c
@@ -324,7 +324,7 @@ unsigned int apic_mmio_access(struct registers *guest_regs,
 			      bool is_write)
 {
 	struct mmio_access access;
-	unsigned long val;
+	u32 val;
 
 	access = mmio_parse(cpu_data, rip, page_table_addr, is_write);
 	if (access.inst_len == 0)
